+++ application/tests/Bootstrap.php
@@ -53,11 +53,22 @@
  *
  * NOTE: If you change these, also change the error_reporting() code below
  */
-// This `if` statemant is needed for @runInSeparateProcess
-if (! defined('ENVIRONMENT'))
+
+// Define testing environment for ci-phpunit-test
+if ( ! defined('ENVIRONMENT'))
 {
+	// The above `if` statement is needed for @runInSeparateProcess
 	define('ENVIRONMENT', 'testing');
 }
+// If you want to change `testing`, you must define `is_testing_env()`.
+//if ( ! function_exists('is_testing_env'))
+//{
+//	// The above `if` statement is needed for @runInSeparateProcess
+//	function is_testing_env()
+//	{
+//		return (ENVIRONMENT === 'unittest');
+//	}
+//}
 
 /*
  *---------------------------------------------------------------