+++ application/tests/Bootstrap.php
@@ -60,15 +60,6 @@ if ( ! defined('ENVIRONMENT'))
 	// The above `if` statement is needed for @runInSeparateProcess
 	define('ENVIRONMENT', 'testing');
 }
-// If you want to change `testing`, you must define `is_testing_env()`.
-//if ( ! function_exists('is_testing_env'))
-//{
-//	// The above `if` statement is needed for @runInSeparateProcess
-//	function is_testing_env()
-//	{
-//		return (ENVIRONMENT === 'unittest');
-//	}
-//}
 
 /*
  *---------------------------------------------------------------