+++ application/tests/_ci_phpunit_test/CIPHPUnitTestCase.php
@@ -456,4 +456,14 @@ class CIPHPUnitTestCase extends TestCase
 		$result = CIPHPUnitTestLogger::didLog($level, $message);
 		$this->assertTrue($result);
 	}
+
+	/**
+	 * Testing Environment or not?
+	 *
+	 * @return bool
+	 */
+	public static function isTestingEnv()
+	{
+		return (ENVIRONMENT === 'testing');
+	}
 }