+++ application/tests/_ci_phpunit_test/patcher/4.x/Patcher/MethodPatcher/NodeVisitor.php
@@ -10,22 +10,39 @@
 
 namespace Kenjis\MonkeyPatch\Patcher\MethodPatcher;
 
+use Kenjis\MonkeyPatch\Patcher\MethodPatcher;
 use PhpParser\Node;
 use PhpParser\Node\Stmt\ClassMethod;
 use PhpParser\NodeVisitorAbstract;
-
-use Kenjis\MonkeyPatch\Patcher\MethodPatcher;
+use PhpParser\ParserFactory;
 
 class NodeVisitor extends NodeVisitorAbstract
 {
 	public function leaveNode(Node $node)
 	{
-		if (! ($node instanceof ClassMethod))
-		{
+		if (! $node instanceof ClassMethod) {
 			return;
 		}
 
-		$pos = $node->getAttribute('startTokenPos');
-		MethodPatcher::$replacement[$pos] = true;
+		$parser = (new ParserFactory())->create(ParserFactory::PREFER_PHP7);
+
+		if ($node->returnType !== null) {
+			if (isset($node->returnType->name) && $node->returnType->name === 'void') {
+				$ast = $parser->parse('<?php ' . MethodPatcher::CODENORET);
+			} elseif (isset($node->returnType->parts) && $node->returnType->parts[0] === 'void') {
+				$ast = $parser->parse('<?php ' . MethodPatcher::CODENORET);
+			} else {
+				$ast = $parser->parse('<?php ' . MethodPatcher::CODE);
+			}
+		} else {
+			$ast = $parser->parse('<?php ' . MethodPatcher::CODE);
+		}
+
+		if ($node->stmts !== null) {
+			array_unshift(
+				$node->stmts,
+				$ast[0]
+			);
+		}
 	}
 }