+++ application/tests/_ci_phpunit_test/ChangeLog.md
@@ -5,7 +5,7 @@
 ### Added
 
 * Add functionality to create mocks on consecutive calls. See [#339](https://github.com/kenjis/ci-phpunit-test/pull/339).
-* Add functionality to change `ENVIRONMENT` constant value `testing`. See [#354](https://github.com/kenjis/ci-phpunit-test/pull/354).
+* Add functionality to change `ENVIRONMENT` constant value `testing`. See [#360](https://github.com/kenjis/ci-phpunit-test/pull/360).
 
 ### Fixed