+++ application/tests/_ci_phpunit_test/ChangeLog.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+* Now you can specify custom test directory. See [#362](https://github.com/kenjis/ci-phpunit-test/pull/362).
 * Add functionality to create mocks on consecutive calls. See [#339](https://github.com/kenjis/ci-phpunit-test/pull/339).
 * Add functionality to change `ENVIRONMENT` constant value `testing`. See [#360](https://github.com/kenjis/ci-phpunit-test/pull/360).